3d tree renderer and animation

+++ ConsoleApplicationRTree/Helpers/Helpers/libhelpers/Structs.h
@@ -39,7 +39,7 @@ namespace Structs {
 		};
 
 		Rgba();
-		Rgba(uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255);
+		Rgba(int r, int g, int b, int a = 255);
 		Rgba(float r, float g, float b, float a = 1.0f);
 	};
